添加程序自定义函数的脚本；修改存储过程代码；修改测试数据

+++ src/main/resources/configs/db/ParkingLot.sql
@@ -46,7 +46,7 @@ create table Account
 create table Bill
 (
 	id bigint primary key auto_increment
-	,userId int 
+	,userId bigint 
 	,parkingLotId int 
 	,accountId bigint
